Reimplemented opposite quiver and opposite path algebra.
The new files opposite.g{d,i} contain operations for finding the
opposites of quivers and path algebras.  This implementation contains
the following improvements over the one in oppdualtrans.gi:

-- The opposites are stored as attributes, so calling OppositeQuiver
   or OppositePathAlgebra several times on the same object produces
   exactly the same result (as compared by IsIdenticalObj).
   Additionally, taking the opposite of the opposite gives the
   original object back.

-- There are operations for finding the opposite of a path in a quiver
   and an element of (a quotient of) a path algebra.

I have also added the following operations which I think should exist:

-- QuiverContainingPath

-- PathAlgebraContainingElement

-- OriginalPathAlgebra: Takes a quotient of a path algebra as argument
   and returns the path algebra it is a quotient of.

Some functions which should have been operations are instead written
as global functions, since there is no filter for testing if a given
object is an element of a path algebra.