Quick fix to make maps from nonzero modules to zero modules work.
There seems to be serious problems with zero modules.  For some
reason, the element of a zero module is completely different from an
element of a nonzero module.

This commit does not really fix the problem (since I do not understand
why it happens), but just makes the homomorphism code work around it.