Fixed some code that didn't work for quivers with no arrows.
The module constructor which only takes information about the action
for each arrow, and not the dimension vector, does not work if the
quiver has no arrows (or if there are isolated vertices).
Nevertheless, it is used several places in the code.

As a quick fix, I wrapped calls to this constructor in a check for
whether the quiver has any arrows.  This fixes the problem for quivers
with no arrows, but not for quivers with arrows and with isolated
vertices.

The correct fix would be to never use such unsafe constructors.