Complexes can automatically detect repetition and change to "repeat" type.
New type "next/repeat" which can be used when specifying complexes.
This works like the type "next" (each differential is produced by a
function which takes the previous differential as argument), but will
automatically convert itself to type "repeat" (all differentials are
given by a finite list which is repeated indefinitely) if it detects
repetition.