New function MakeQPADocumentation for building documentation.
For some users, it may not be convenient to run GAP from the command
line, and then it is difficult to use the MakeQPADoc.gi (which relies
on being run with the directory the file is in as current directory).
It is now possible to build the documentation by calling the function
MakeQPADocumentation from within GAP instead.