Turn IsDirectSumOfModules into a pure filter
IsDirectSumOfModules is not a property or attribute that can be
set to true or false due to a later computation; rather, it can only
be set when creating an object via a special constructor.

That suggests it should simply be a filter, not an attribute.

This makes qpa compatible with this upcoming GAP change:
<https://github.com/gap-system/gap/pull/2732>