Enhance complex constructor: discover zero complexes
In certain cases it is easy to see from the data used to construct a
complex that the complex must consist of only zero objects and zero
maps.  I tried to make the complex constructor aware of some such
cases and mark the resulting complex as an instance of IsZeroComplex.