Undo change from commit bd09e5b (Fix bug in TauOfComplex).
That commit was made in order to fix issue #35.  I believe the change
made there is unnecessary, and that the issue is better fixed by the
three commits leading up to this one (960f88f, c090710 and 8df7e3b),
which fix bugs in complex.gi while leaving derivedcat.gi unchanged.